fix: 修正了测试中出现的一些错误，主要集中在interface的数字是float64和log遗漏了userId

+++ biz/application/service/user.go
@@ -65,7 +65,7 @@ func (u *UserService) SignUp(ctx context.Context, req *show.SignUpReq) (*show.Si
 	return &show.SignUpResp{
 		Id:           userId,
 		AccessToken:  authorization,
-		AccessExpire: signUpResponse["accessExpire"].(int64),
+		AccessExpire: int64(signUpResponse["accessExpire"].(float64)),
 	}, nil
 }
 
@@ -80,7 +80,7 @@ func (u *UserService) SignIn(ctx context.Context, req *show.SignInReq) (*show.Si
 	return &show.SignInResp{
 		Id:           signInResponse["userId"].(string),
 		AccessToken:  signInResponse["accessToken"].(string),
-		AccessExpire: signInResponse["accessExpire"].(int64),
+		AccessExpire: int64(signInResponse["accessExpire"].(float64)),
 	}, nil
 }
